# apifox/openapi.yaml
openapi: 3.0.3
info:
  title: dateOrFriends API
  version: 0.0.1-SNAPSHOT
  description: |
    dateOrFriends 后端接口文档（可直接导入 Apifox：导入 -> OpenAPI）。

    统一返回结构：{ code, message, data }
servers:
  - url: http://localhost:8888
security:
  - bearerAuth: []
...
paths:
...
  /admin/notice/list:
    get:
      tags: [Admin]
      summary: 公告管理-列表
      description: 管理端分页查询公告（ADMIN/SUPER_ADMIN）
      security:
        - bearerAuth: []
      parameters:
        - name: keyword
          in: query
          required: false
          schema:
            type: string
          description: 标题/内容关键词（全文索引）
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [DRAFT, PUBLISHED, OFFLINE]
        - name: level
          in: query
          required: false
          schema:
            type: string
            enum: [NORMAL, IMPORTANT]
        - name: pinned
          in: query
          required: false
          schema:
            type: integer
            format: int32
            enum: [0, 1]
        - name: publishFrom
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: publishTo
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageNotice'
        '401':
          description: 未登录或令牌无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
  /admin/notice/create:
    post:
      tags: [Admin]
      summary: 公告管理-创建
      description: 创建公告（默认草稿）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNoticeRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseNotice'
        '400':
          description: 参数不合法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '401':
          description: 未登录或令牌无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
  /admin/notice/update:
    post:
      tags: [Admin]
      summary: 公告管理-更新
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNoticeRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseNotice'
        '400':
          description: 参数不合法/公告不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '401':
          description: 未登录或令牌无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
  /admin/notice/publish:
    post:
      tags: [Admin]
      summary: 公告管理-发布
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseNotice'
        '400':
          description: 公告不存在/操作失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '401':
          description: 未登录或令牌无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
  /admin/notice/offline:
    post:
      tags: [Admin]
      summary: 公告管理-下线
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseNotice'
        '400':
          description: 公告不存在/操作失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '401':
          description: 未登录或令牌无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
  /admin/notice/delete:
    post:
      tags: [Admin]
      summary: 公告管理-删除
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '400':
          description: 公告不存在/操作失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '401':
          description: 未登录或令牌无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
  /admin/notice/{id}:
    get:
      tags: [Admin]
      summary: 公告管理-详情
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseNotice'
        '400':
          description: 参数不合法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '401':
          description: 未登录或令牌无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
  /admin/notice/logs:
    get:
      tags: [Admin]
      summary: 公告管理-操作日志
      security:
        - bearerAuth: []
      parameters:
        - name: noticeId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: actorId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: actionType
          in: query
          required: false
          schema:
            type: string
            enum: [CREATE, UPDATE, PUBLISH, OFFLINE, DELETE]
        - name: fromTime
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: toTime
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageNoticeLog'
        '401':
          description: 未登录或令牌无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
  /admin/notice/logs/export:
    get:
      tags: [Admin]
      summary: 公告管理-导出操作日志
      security:
        - bearerAuth: []
      parameters:
        - name: noticeId
          in: query
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: CSV 文件
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '401':
          description: 未登录或令牌无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
...
  /user/notices/{id}:
    get:
      tags: [User]
      summary: 用户端-公告详情（并标记已读）
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUserNotice'
        '400':
          description: 参数不合法/公告不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '401':
          description: 未登录或令牌无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponseVoid:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          format: int32
          description: 0=成功；非0=失败
          example: 0
        message:
          type: string
          example: ok
        data:
          nullable: true
...
    ApiResponseError:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          format: int32
          description: 业务错误码（例如 1000 参数不合法、1001 未授权、1002 无权限、2000 账号不存在 等）
          example: 1001
        message:
          type: string
          description: 错误信息
          example: 未登录或令牌无效
        data:
          nullable: true
...
    Notice:
      type: object
      required: [id, title, content, level, status, pinned, createdBy, updatedBy, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        content:
          type: string
        level:
          type: string
          enum: [NORMAL, IMPORTANT]
        status:
          type: string
          enum: [DRAFT, PUBLISHED, OFFLINE]
        pinned:
          type: boolean
        publishAt:
          type: string
          format: date-time
          nullable: true
        expireAt:
          type: string
          format: date-time
          nullable: true
        createdBy:
          type: integer
          format: int64
        updatedBy:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateNoticeRequest:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          maxLength: 128
        content:
          type: string
        level:
          type: string
          enum: [NORMAL, IMPORTANT]
          default: NORMAL
        pinned:
          type: boolean
          default: false
        expireAt:
          type: string
          format: date-time
          nullable: true
    UpdateNoticeRequest:
      type: object
      required: [id, title, content]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
          maxLength: 128
        content:
          type: string
        level:
          type: string
          enum: [NORMAL, IMPORTANT]
          default: NORMAL
        pinned:
          type: boolean
          default: false
        expireAt:
          type: string
          format: date-time
          nullable: true
    IdRequest:
      type: object
      required: [id]
      properties:
        id:
          type: integer
          format: int64
    NoticeLog:
      type: object
      required: [id, noticeId, actionType, createdAt]
      properties:
        id:
          type: integer
          format: int64
        noticeId:
          type: integer
          format: int64
        actorId:
          type: integer
          format: int64
          nullable: true
        actorRole:
          type: string
          nullable: true
        actionType:
          type: string
          enum: [CREATE, UPDATE, PUBLISH, OFFLINE, DELETE]
        ip:
          type: string
          nullable: true
        detail:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    PageNotice:
      type: object
      required: [total, page, size, items]
      properties:
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: '#/components/schemas/Notice'
    ApiResponseNotice:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          format: int32
          example: 0
        message:
          type: string
          example: ok
        data:
          $ref: '#/components/schemas/Notice'
    ApiResponsePageNotice:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          format: int32
          example: 0
        message:
          type: string
          example: ok
        data:
          $ref: '#/components/schemas/PageNotice'
    PageNoticeLog:
      type: object
      required: [total, page, size, items]
      properties:
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: '#/components/schemas/NoticeLog'
    ApiResponsePageNoticeLog:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          format: int32
          example: 0
        message:
          type: string
          example: ok
        data:
          $ref: '#/components/schemas/PageNoticeLog'
    UserNotice:
      type: object
      required: [id, title, content, level, pinned, read]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        content:
          type: string
        level:
          type: string
        pinned:
          type: boolean
        publishAt:
          type: string
          format: date-time
          nullable: true
        expireAt:
          type: string
          format: date-time
          nullable: true
        read:
          type: boolean
...
    ApiResponseUserNotice:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          format: int32
          example: 0
        message:
          type: string
          example: ok
        data:
          $ref: '#/components/schemas/UserNotice'
...